feat: 구조 재변경
